Acquire GITHUB_AUTH_TOKEN from OS environment
In preparation for a forth-coming Pelican plugin I am developing,
I needed to add a GitHub auth token to my Pelican settings. However,
it would not be a great idea to expose this token in the settings file.
Thus, I am now attempting to acquire the value from the OS environment
setting 'GITHUB_AUTH_TOKEN' -- if the value doesn't exist, a
NotImplementedError will be raised that presents a message on how to
continue.

Furthermore, the Makefile was updated to call the setup_env.sh script,
a not-to-be-committed-to-git file that is essentially:

export GITHUB_AUTH_TOKEN='xyz....'